@@ -12,7 +12,7 @@
     "fs": "^0.0.1-security",
     "http": "^0.0.1-security",
     "https": "^1.0.0",
-    "node-fetch": "^3.1.0",
+    "node-fetch": "^3.2.0",
     "sass-loader": "^10.2.0",
     "saxon-js": "^2.3.0",
     "stream": "^0.0.2",
@@ -23,17 +23,16 @@
     "zlib": "^1.0.5"
   },
   "devDependencies": {
-    "@vue/cli-plugin-eslint": "~4.5.0",
-    "@vue/cli-plugin-router": "~4.5.0",
-    "@vue/cli-service": "~4.5.0",
+    "@vue/cli-plugin-eslint": "^5.0.1",
+    "@vue/cli-plugin-router": "^5.0.1",
+    "@vue/cli-service": "^5.0.1",
     "@vue/eslint-config-prettier": "^6.0.0",
-    "acdh-fundament-vuetify": "^1.1.1",
     "arche-api": "^1.2.1",
     "bootstrap": "^5.1.3",
     "bootstrap-vue": "^2.21.2",
-    "eslint": "^6.7.2",
+    "eslint": "^8.10.0",
     "eslint-plugin-prettier": "^3.3.1",
-    "eslint-plugin-vue": "^6.2.2",
+    "eslint-plugin-vue": "^8.5.0",
     "prettier": "^2.2.1",
     "vue": "^2.6.14",
     "vue-template-compiler": "^2.6.11"
