@@ -81,6 +81,6 @@
 
 <script>
 export default {
-  name: "Footer",
+  name: "footer-item",
 };
 </script>
