@@ -7,12 +7,12 @@ Vue.use(VueRouter);
 const routes = [
   {
     path: "/",
-    name: "Home",
+    name: "home-item",
     component: Home,
   },
   {
-    path: "/about",
-    name: "About",
+    path: "/about-item",
+    name: "about-item,",
     // route level code-splitting
     // this generates a separate chunk (about.[hash].js) for this route
     // which is lazy-loaded when the route is visited.
@@ -22,7 +22,7 @@ const routes = [
   },
   {
     path: "/started",
-    name: "Started",
+    name: "started-item",
     component: () =>
       import(/* webpackChunkName: "about" */ "../views/Started.vue"),
   },
