@@ -1,5 +1,17 @@
 <template>
-  <div class="content">
-    <h1>This is an about page</h1>
+  <div class="About">
+    <About></About>
   </div>
 </template>
+
+<script>
+// @ is an alias to /src
+import About from "@/components/About.vue";
+
+export default {
+  name: "about-item",
+  components: {
+    About,
+  },
+};
+</script>
